+++ VulkanEngine/Vermillion/src/core/vulkan/render_passes/lightfield/disparity_renderpass.hpp
@@ -318,7 +318,7 @@ private:
 
 		// Finally, create actual render pipeline
 		vk::GraphicsPipelineCreateInfo graphicsPipelineInfo = vk::GraphicsPipelineCreateInfo()
-			.setStageCount(shaderStages.size())
+			.setStageCount((uint32_t)shaderStages.size())
 			.setPStages(shaderStages.data())
 			// fixed-function stages
 			.setPVertexInputState(&vertexInputInfo)
